Allow to use dbus on windows
